fixed lst mint error

+++ programs/stakrr/src/lib.rs
@@ -12,7 +12,7 @@ declare_id!("5GZcSx7NBYH6HEFiqGfSNJZVWz7qKvWaXR5Esk8UXxJU");
 pub mod stakrr {
     use super::*;
 
-    pub fn initialize(ctx: Context<InitializePool>, rate_per_second: u64) -> Result<()> {
+    pub fn initialize(ctx: Context<InitializePool>, rate_per_second: f64) -> Result<()> {
         handle_initialize_pool(ctx, rate_per_second)
     }
 
@@ -32,13 +32,14 @@ pub mod stakrr {
         handle_fund_reward_pool(ctx, amount)
     }
 
-    pub fn update_reward_rate(ctx: Context<UpdateRewardRate>, new_rate: u64) -> Result<()> {
+    pub fn update_reward_rate(ctx: Context<UpdateRewardRate>, new_rate: f64) -> Result<()> {
         handle_update_reward_rate(ctx, new_rate)
     }
 
     pub fn pause(ctx: Context<Pause>) -> Result<()> {
         handle_pause(ctx)
     }
+
     pub fn unpause(ctx: Context<Unpause>) -> Result<()> {
         handle_unpause(ctx)
     }
